changes path when turrets and walls are placed
When a turret or wall is placed onto the level, the outside enemies find
a path to the tower using dijkstra. For now if the enemy has already
been spawned it'll stay on its original course, the newly spawned
enemies outside follow the new path. If all paths become blocked the
enemies use the old path for now.